Remove NavigationLink workaround in ContactsView
This workaround mainly existed to give us the NavigationLink chevron on
the trailing edge of each contact entry, without needing to do any
manual styling.

However, on iOS 18.0, nesting the buttons and NavigationLink in this
way causes the inner button not to be tappable. So, let's remove this
workaround and rely on the default styling.