Remove stale promises from the DB on app startup
Before this commit, promises were only removed from the DB once they
were consumed.

Therefore, it was possible to end up with "stale" promises lingering in
the promise table of the DB, if the promise was created and persisted,
but the app got closed before they could be consumed.

These promises would linger forever because there was no way to consume
them anymore (resolvers map was empty) and would therefore just take up
space in the table for no reason.

This commit fixes this by truncating the promise table on each restart
of the application. We can truncate the whole table as we know that all
of the promises still in the DB at app start-up are unconsumable, as
the resolvers map will be empty at this point.