Remove duplicate hideLoadingOverlay
There are overloaded instances of showPromisingLoadingOverlay - some of
which automatically hide the overlay once the promise is resolved, and
some which do not and expect the caller to hide the overlay.

In ContactDetails.swift, we only call the instances that hide the
overlay themselves on consumption of the promise.

Therefore, this extra call to hideLoadingOverlay is redundant - the
promise overlay class itself should take care of this.