Use framework import style for monalxmpp headers
Apple would like us to use "framework style" (<>) imports for framework
headers. Framework-style headers help ensure that the framework is
fully contained; if a framework was to use a local header, there may be
compilation errors if the framework is used out of the current context.

For consistency, use framework-style imports for monalxmpp imports
across the codebase, so that the developer does not have to think
whether they are in a framework header not when importing the file.